Add base fragment for elective `Subject` handling
