+++ app/src/main/java/com/senex/timetable/presentation/ui/subject/varied/BaseVariedSubjectFragment.kt
@@ -1,0 +1,39 @@
+package com.senex.timetable.presentation.ui.subject.varied
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.viewbinding.ViewBinding
+import com.senex.timetable.presentation.common.inflateBinding
+import dagger.android.support.DaggerFragment
+
+abstract class BaseVariedSubjectFragment<T : ViewBinding> : DaggerFragment() {
+    private var _binding: T? = null
+    protected val binding
+        get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ) = inflateBinding(getBindingInflater, layoutInflater, container) {
+        _binding = it
+    }
+
+    protected abstract val getBindingInflater: (LayoutInflater, ViewGroup?, Boolean) -> T
+
+    override fun onViewCreated(
+        view: View,
+        savedInstanceState: Bundle?,
+    ) = onViewCreatedCallback(binding)
+
+    protected abstract val onViewCreatedCallback: T.() -> Unit
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
+
+
