+++ app/src/main/java/com/senex/timetable/presentation/ui/subject/varied/NewElectiveSubjectViewModel.kt
@@ -4,7 +4,6 @@ import com.senex.timetable.domain.usecase.subject.GetAllByElectiveSubjectId
 import com.senex.timetable.domain.usecase.subject.elective.GetElectiveSubject
 import com.senex.timetable.domain.usecase.subject.elective.HideElectiveSubject
 import com.senex.timetable.domain.usecase.subject.elective.ShowElectiveSubject
-import com.senex.timetable.presentation.ui.subject.elective.ElectiveSubjectViewModel
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
